Prerelease 0.01
I've been spending alot of time on and off creating a modular "framework" that I plan to use for future projects and games.

There are a lot of changes and improvements that need to be done with this codebase for me to call it complete.

Some of the current unfinished features include custom console commands, creating a match from the DLL,  ImGui/KieroHook menu, possibly an ALL.Net proxy (scrapped),  and a rewrite of the hooking structure.