Rewrote the hooking logic.
I hated using the old logic, it made sense at first but then turned into abysmal garbage that even I didn't want to use.

I've learned a lot of new concepts in C++ since I first started working on this project.

I implemented what I learned with structs to make creating hooks modular and user-friendly, which was the original vision.

The actual game still functions the exact same so there isn't a need for me to precompile a DLL and have everyone update to using it.