Major changes/cleanup
At this point we could call this an entire rewrite of a lot of major components, first off I will say that the assembly code/proxy hook idea currently does not work and I don't know when I will get the motivation to try getting it to work again, I'm still learning.

Anyways, I have rewritten UFunctions, added separate logic for hooks, offsets, changed from MinHook Library to the source, added a completely new proxy library I have been cooking up.

UE4 Librarian or any DLL sideloader/injector is no longer needed, as the DLL itself is it's own proxy and injector.

Everything is still an unfinished work in progress, but I haven't made a commit in a while so I'm doing that now.

I added some ideas/scrap code such as custom commands/internal commands, both of which currently also do not work as of the moment, plenty of test code.

Basically the code comes as-is, it's a dev mess.