Enabling Loose File Loading [LFL] with Pakfiles
With native Unreal Engine as far as I'm aware, there is no option to load pak files and loose files at the same time. It's one or the other, which is annoying when people want to mod the game.

So I'm working on writing some logic to allow for LFL while still using the core game as a pak file, not only to save space and keep organized but also to prevent inexperienced users from accidentally vital parts of the game.

Currently I've hooked IsNonPakFilenameAllowed and FindFileInPakFiles, I'm pretty sure that FindFileInPakFiles is the function I actually want to use for the logic.
It's unfinished but my concept works perfectly.