COMPLETELY UNUSABLE
I spent hours working on this unfinished rewrite of a lot of different systems, the end result will be a really clean template-able codebase.
But for now I just decided it would be a good time to make a commit incase something bad were to happen and I lost all the work I made with actual clarity.