Working on finishing up the rewrite
Still a bit of ways away but I think I finished the Hooks wrapper class.

I haven't used or tested it, infact I passed out while working on it so I'm not entirely sure where to go from here because my mind is a random number generator which sometimes wants to work on it and other times it's a foggy mess where I want to just go to sleep, it's hot in here.