Compilable, not usable.
I'm tired, today was a success though as I've really made a lot of progress in one go.

I've had to rework some things because of said tiredness but now it compiles and kinda works.

As in I was able to successfully hook Browse and not return the original call because the logic is unfinished, first time I have ever seen the game in limbo but with the UConsole unlocked.