Reimplementing Basic Features
I've implemented a prototype BytePatching module/library that effectively works at killing ResetPP and StartConsumePP.

I'm working on adding a new CustomCommands module that will actually work efficiently and just make sense, unfortunately today or right now is not a good time to work on that logic, instead I've implemented the basic features needed to get the game to launch safely on (hopefully) everyone's computers without a random nullptr crash/etc.

I've been on and off of this project and I'm hoping to finally get something going, this commit is still quite small and minimal but I'm planning to really dig in and start the grind in the upcoming days.