Testing bInputDebugMode
Improved the logic of my Pointers module, added Login hooks and a little miniature test for a pause menu.

Also fixed a compile problem that I forgot existed due to the .gitignore excluding the Resource folder.

This is a few days late, I simply forgot to commit it, I'm going to start working on a properly usable PC port release right now.