I spent 1 hour and 16 minutes obsessing for no reason
This is why I take forever to do literally anything, because I get distracted with the stupidest and most pointless intricate details like trying my best to only use a singular byte (00000000) and with the help of ChatGPT to explain to me how to do bitshifting operations I was able to make this abomination.

I am going to clean it up, I am simply committing this so I don't lose all of the work I did with different tests/etc.

I completely understand the logic and have zero idea what any of this means at the same time, the idea was that since the iterator only goes up to 3 and 00000011 is the biggest it goes, I could cram the rest of the logic/flags for the loop into the other bits.

The logic works and the code runs while ONLY logging the referred files/strings in the LocalDirectory, the problem is I don't know if me manually doing this is either computationally slower or faster since...-computers, but I was too obsessed over learning how to do bitshifting so here we go, maybe this code goes in the garbage bin later but the concept works perfectly and manifests.