Full modding support added!
I originally had planned to add a feature that would loose file load ONLY if your loose file had an _P before the extension, similar to how Unreal does it with files.

While my logic in itself absolutely worked, the files would not load since the game was not expecting the filename to have _P inside, so if I really wanted to nail that logic down I would have to dig in and make everything more of a mess than need be.

It was just an idea to idiot-proof accidental mods/files from being loaded that are unwanted/unneeded, maybe it will be a requirement in the future but for now it works perfectly fine without it.

I had to also hook IsNonPakFilenameAllowed to allow for *adding* mods, FindFileInPakFiles only allows you to *replace* internal files with loose files, they work together.

I'm going to sleep now but I definitely plan to put out a new PC Port release real soon, it's been overdue and one of the biggest reasons was that mod support was annoying to deal with, in addition to the bDebugInputMode being bugged and online connectivity not having been finished.

I am going to work on fixing bDebugInputMode within the coming days.