Killed the HideCursor function FINALLY
For countless months I had been mistaken, I thought that the function was there to *show* the cursor, after I looked into the pseudocode with a currently better understanding I realized it was simply garbage that needed to be removed.

Now your mouse can be displayed anywhere if Unreal Engine's context allows for it, previously it would be mysteriously blocked even if Unreal Engine was supposed to show it, now we know why.

I also learned and realized that my original SDK dump technically did not have everything, that is where v2 comes from, I dumped the SDK while ingame (specifically AJBSimpleMatch_P) instead of dumping it unpatched somewhere else (probably the arcade check level)

 I put it in a separate folder to differentiate the two, especially if it turns out that these extra headers/files aren't so important after-all, I also dumped that SDK with my game already patched, so a few of my mods (enhancements) have contaminated the SDK.

On a sidenote, I fixed a design flaw with my Pointers module, it was returning a probable dangling pointer.