Working on custom commands and internal commands.
I have finally gotten back to this *nightmare* and my brain is finally in the position to get it working, and indeed it does.

I've only written one part of it (I am sleep deprived) but I will continue to work on the implementation.

I've wanted custom commands for a long time, and I've made numerous versions of them, but they were either hardcoded or badly written/inefficient.

I will 100% be changing up the CommandStructure, but the main idea is that I will be designing a dynamic custom command library/module that will hopefully work in a similar/indistinguishable way to native Unreal console commands.

The reason I started working on this (instead of anything else with the game) is that I'm working on making a custom menu/frontend for the game, and the only logical way that I will be able to properly communicate with the DLL from a modded widget blueprint is by executing different console commands and parsing them through my DLL.

I've also been needing a dynamic/universal custom command parser for quite a long time now... so this kills two birds with one stone.