Did a lot of testing, continued on custom commands, and made use of SpawnActor
I spent a lot of hours going insane over scrap code, worked on custom command parsing, lost my mind again, and finally took a break.

I'm very lost on how I will be designing this system, since I plan to get a release soon I will not be worrying too much about it.

If my extreme obsessive compulsive behavior isn't making much sense in the context of this game, it's because I'm designing most of these systems to be dependencies, so I can integrate them in plenty of other games.

This game is my version of *testing in prod*, since it's currently my biggest project, atleast biggest in the sense of the amount of people running my code on their computers.

I finally figured out how to natively play the backround music from the game's audio system, but in my custom maps the WWiseManager instance ceases to exist, so I have to manually spawn it.