Genuine insanity from like 3 weeks ago
There isn't much in this commit besides a few changes, I didn't commit it since I was just debugging to try and figure out how the CloseTime function was still calling and still preventing the frontend from functioning.
I figured out a little bit on that day but I didn't get very far, and now I don't even remember what any of this means because I didn't do a good job naming them, this commit is for possible use/reference in the future but will be wiped out in the next commit.