PC Port v28 for 1.28.02
This was a complete overhaul.

I've added dummy libraries to replace some arcade-managing libraries that I honestly don't trust, the one that got my attention is AMActivator, it has functions designed to activate and deactivate bitlocker, and while it is unlikely/improbable I would rather not take any risks for some low probability of it affecting someone.

I kinda cleaned the codebase, moved the SDK in a folder to debloat the main directory.

I had to change all of the offsets manually and generate the SDK for this new build, but it wasn't bad at all.

I've changed the design of my PC Port to be more dynamic and modular. The DLL controls most if not all of the game's configs rather than the PC Port pak, I've isolated the pure raw game into a pakfile called AJB-CORE.pak

I added a lot more logic towards the game to allow for ease of use, a miniature SDK designed for this game specifically.

The DLL is now a little over 100kb larger due to the ico image I have embedded into it, this adds the image to the game's process for polish, instead of using a blank Unreal Engine logo.

The game is fully mod supported with loose file loading, allowing for old mods and new ones. Infact my PC Port release is currently using the controller inputs as a loose file to allow for people to easily configure it.

Multiplayer works, but it's still not very user-friendly, I'm planning to add a lot more support in the near future, singleplayer and custom maps work perfectly fine though.

A lot of other improvements/changes have been added, I can't list them all from the top of my head but this has been a much needed improvement, a MASSIVE improvement.