Recreating my own wheel
I'm working on my CArray class which is essentially a std::vector but not unreadable bloat, I'm reinventing the wheel because I want to learn how the wheel was made.

Afterwards I can subclass it and make a CStr class for resizable CStrings, and finally I can use those to recreate my logger, which is probably my oldest
and most outdated library that I have in my codebase.

Once I remove strings and vectors entirely with my own implementation The size of the DLL (while small) should decrease significantly, because in my opinion its currently not small enough.