Custom UI integration prototype v2
Most of the work was done in the blueprint, I am way too tired to write a commit note, I spent way too long trying to create the blueprint class (which I finally got working) without using a level blueprint which was my old method.
But I also wanted the class to be persistent, to not die everytime I switch levels, currently that is an issue that will need to be solved alongside other things.

This codebase needs some serious cleaning up and I need to finish my libraries.