Messed with way too many things and got distracted a bunch
This commit probably contains almost 1000 lines of code, with most of it being scrap/debug/testing stuff, I have not committed in a long time.

The summary of what I accomplished here is that I managed to force load a custom cursor blueprint into the GConfig (or something), discovered a cool bug/feature that removes the widget alignment enforcement (it makes the game truly look like a PC game), and discovered an internal debug menu that I cannot currently get working.

I'm planning to clean this code up soon but this month is going to be a miserable and long grind, and I have other projects to attend to...

I have so much unfinished functionality scattered in this codebase currently, selecting characters works/but selecting skins is broken, I got bored and stopped working on a UObject flag filter because it wasn't working, and then I did who knows what else, let this commit speak for itself I guess.